LayoutTests/fast/loader/data-url-encoding-svg.html

<meta charset="koi8-r">
<body onload="test()">
<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();

function test() {
    document.getElementById("result").innerHTML = "1. " + (window.frames[0].document.characterSet == "windows-1251" ? "SUCCESS" : "FAILURE: " + window.frames[0].document.characterSet) + ".\r";
    document.getElementById("result").innerHTML += "2. " + (window.frames[1].document.documentElement.getAttribute("bar") == "SUóóåSS" ? "SUCCESS" : "FAILURE: " + window.frames[1].document.documentElement.getAttribute("bar")) + ".\r";
    document.getElementById("result").innerHTML += "3. " + (window.frames[2].document.characterSet == "UTF-8" ? "SUCCESS" : "FAILURE: " + window.frames[2].document.characterSet)
        + ". The result is currently wrong due to &lt;rdar://problem/5845875> REGRESSION: -[NSURLResponse textEncoding] incorrectly reports us-ascii encoding for data: URLs.\r";
    document.getElementById("result").innerHTML += "4. " + (window.frames[3].document.documentElement.getAttribute("bar") == "SUóóåSS" ? "SUCCESS" : "FAILURE: " + window.frames[3].document.documentElement.getAttribute("bar"))
        + ". The result is currently wrong due to &lt;rdar://problem/5845875> REGRESSION: -[NSURLResponse textEncoding] incorrectly reports us-ascii encoding for data: URLs.\r";
}
</script>
<p>Test that non-ASCII data: URLs are handled correctly.</p>
<pre id=result></pre>
<iframe src="data:image/svg+xml;charset=windows-1251,<svg xmlns=%22http://www.w3.org/2000/svg%22></svg>"></iframe>
<iframe src="data:image/svg+xml;charset=utf-8,<svg xmlns=%22http://www.w3.org/2000/svg%22 foo='#' bar='SUóóåSS'></svg>"></iframe>
<iframe src="data:image/svg+xml,<svg xmlns=%22http://www.w3.org/2000/svg%22></svg>"></iframe>
<iframe src="data:image/svg+xml,<svg xmlns=%22http://www.w3.org/2000/svg%22 foobar='#' bar='SUóóåSS'></svg>"></iframe>
</body>
