LayoutTests/fast/loader/data-url-encoding-svg-expected.txt

Test that non-ASCII data: URLs are handled correctly.

1. SUCCESS.
2. SUCCESS.
3. FAILURE: US-ASCII. The result is currently wrong due to <rdar://problem/5845875> REGRESSION: -[NSURLResponse textEncoding] incorrectly reports us-ascii encoding for data: URLs.
4. FAILURE: SUССЕSS. The result is currently wrong due to <rdar://problem/5845875> REGRESSION: -[NSURLResponse textEncoding] incorrectly reports us-ascii encoding for data: URLs.
   
